Remove `CurrentSpanPerThread`
It should eventually live in the `tokio-trace-subscriber` crate.
