+++ tokio-trace/src/subscriber.rs
@@ -1,51 +1,6 @@
 //! Collects and records trace data.
 pub use tokio_trace_core::subscriber::*;
 
-use std::{cell::RefCell, default::Default, thread};
-use Id;
-
-/// Tracks the currently executing span on a per-thread basis.
-///
-/// This is intended for use by `Subscriber` implementations.
-#[derive(Clone)]
-pub struct CurrentSpanPerThread {
-    current: &'static thread::LocalKey<RefCell<Vec<Id>>>,
-}
-
-impl CurrentSpanPerThread {
-    pub fn new() -> Self {
-        Self::default()
-    }
-
-    /// Returns the [`Id`](::Id) of the span in which the current thread is
-    /// executing, or `None` if it is not inside of a span.
-    pub fn id(&self) -> Option<Id> {
-        self.current
-            .with(|current| current.borrow().last().cloned())
-    }
-
-    pub fn enter(&self, span: Id) {
-        self.current.with(|current| {
-            current.borrow_mut().push(span);
-        })
-    }
-
-    pub fn exit(&self) {
-        self.current.with(|current| {
-            let _ = current.borrow_mut().pop();
-        })
-    }
-}
-
-impl Default for CurrentSpanPerThread {
-    fn default() -> Self {
-        thread_local! {
-            static CURRENT: RefCell<Vec<Id>> = RefCell::new(vec![]);
-        };
-        Self { current: &CURRENT }
-    }
-}
-
 /// Sets this dispatch as the default for the duration of a closure.
 ///
 /// The default dispatcher is used when creating a new [`Span`] or
